Cleanup project  structure for contract ABI
